copyright: Ban disassembly of userspace gfx blobs
I recognize disassembly on the kernel side is a fact of life, and kernel
device drivers are rarely higher level than bitbanging anyway. However
for userspace, I feel strongly that disassembly/decompilation is
unnecessary and especially risky. As such I do not myself or any Asahi
Mesa contributor to be exposed to potentially tainting disassembly.

As a consequence of the new rule, only #asahi-re is allowed for this
discussion, #asahi-gfx must remain clean. The only corner case might be
disassembly of the AGX kext; I would prefer that be #asahi-re so the
segment of the project I oversee (in #asahi-gfx and primarily Mesa)
remains clean.

Signed-off-by: Alyssa Rosenzweig <alyssa@rosenzweig.io>